+++ packages/medusa/src/api/routes/admin/customers/update-customer.ts
@@ -159,6 +159,9 @@ class Group {
  *   metadata:
  *     description: An optional set of key-value pairs to hold additional information.
  *     type: object
+ *     externalDocs:
+ *       description: "Learn about the metadata attribute, and how to delete and update it."
+ *       url: "https://docs.medusajs.com/development/entities/overview#metadata-attribute"
  */
 export class AdminPostCustomersCustomerReq {
   @IsEmail()