+++ packages/medusa/src/api/routes/admin/discounts/create-dynamic-code.ts
@@ -112,6 +112,9 @@ export default async (req: Request, res: Response) => {
  *   metadata:
  *     type: object
  *     description: An optional set of key-value pairs to hold additional information.
+ *     externalDocs:
+ *       description: "Learn about the metadata attribute, and how to delete and update it."
+ *       url: "https://docs.medusajs.com/development/entities/overview#metadata-attribute"
  */
 export class AdminPostDiscountsDiscountDynamicCodesReq {
   @IsString()