+++ packages/medusa/src/api/routes/admin/draft-orders/update-line-item.ts
@@ -162,6 +162,9 @@ export default async (req, res) => {
  *   metadata:
  *     description: The optional key-value map with additional details about the Line Item.
  *     type: object
+ *     externalDocs:
+ *       description: "Learn about the metadata attribute, and how to delete and update it."
+ *       url: "https://docs.medusajs.com/development/entities/overview#metadata-attribute"
  */
 export class AdminPostDraftOrdersDraftOrderLineItemsItemReq {
   @IsString()