+++ packages/medusa/src/api/routes/admin/products/update-product.ts
@@ -492,6 +492,9 @@ class ProductVariantReq {
  *         metadata:
  *           description: An optional set of key-value pairs with additional information.
  *           type: object
+ *           externalDocs:
+ *             description: "Learn about the metadata attribute, and how to delete and update it."
+ *             url: "https://docs.medusajs.com/development/entities/overview#metadata-attribute"
  *         prices:
  *           type: array
  *           description: An array of product variant prices. A product variant can have different prices for each region or currency code.
@@ -563,6 +566,9 @@ class ProductVariantReq {
  *   metadata:
  *     description: An optional set of key-value pairs with additional information.
  *     type: object
+ *     externalDocs:
+ *       description: "Learn about the metadata attribute, and how to delete and update it."
+ *       url: "https://docs.medusajs.com/development/entities/overview#metadata-attribute"
  */
 export class AdminPostProductsProductReq {
   @IsString()