+++ packages/medusa/src/api/routes/store/customers/update-customer.ts
@@ -116,6 +116,9 @@ export default async (req, res) => {
  *   metadata:
  *     description: "Additional custom data about the customer."
  *     type: object
+ *     externalDocs:
+ *       description: "Learn about the metadata attribute, and how to delete and update it."
+ *       url: "https://docs.medusajs.com/development/entities/overview#metadata-attribute"
  */
 export class StorePostCustomersCustomerReq {
   @IsOptional()