+++ packages/product/src/types/services/index.ts
@@ -1,2 +1,3 @@
+export * as ProductCategoryServiceTypes from "./product-category"
 export * as ProductServiceTypes from "./product"
 export * as ProductVariantServiceTypes from "./product-variant"