+++ packages/oas/oas-github-ci/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/oas-github-ci
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/medusa-oas-cli@0.2.13
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/medusa-oas-cli@0.2.12
+
 ## 1.0.11
 
 ### Patch Changes